requirements.txt
flask>=2.2.0
pandas>=2.0.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
flask-cors>=3.0.10
numpy>=1.24.0
